Split out listener code from html module
